+++ spec/mongoid/criterion/modifiable_spec.rb
@@ -188,4 +188,222 @@ describe Mongoid::Criterion::Modifiable do
       end
     end
   end
+
+  describe "first_or_create" do
+
+    let!(:band) do
+      Band.create(name: "Depeche Mode")
+    end
+
+    context "when the document is found" do
+
+      let(:found) do
+        Band.where(name: "Depeche Mode").first_or_create
+      end
+
+      it "returns the document" do
+        expect(found).to eq(band)
+      end
+    end
+
+    context "when the document is not found" do
+
+      context "when attributes are provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create(origin: "Essex")
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+
+        it "sets the additional attributes" do
+          expect(document.origin).to eq("Essex")
+        end
+      end
+
+      context "when attributes are not provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+      end
+
+      context "when a block is provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create do |doc|
+            doc.active = false
+          end
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+
+        it "yields to the block" do
+          expect(document.active).to be_false
+        end
+      end
+
+      context "when the criteria is complex" do
+
+        context "when the document is not found" do
+
+          let(:document) do
+            Band.in(name: [ "New Order" ]).first_or_create(active: false)
+          end
+
+          it "returns a new document" do
+            expect(document.active).to be_false
+          end
+
+          it "returns a persisted document" do
+            expect(document).to be_persisted
+          end
+        end
+      end
+    end
+  end
+
+  describe "first_or_create!" do
+
+    context "when validation fails on the new document" do
+
+      it "raises an error" do
+        expect {
+          Account.where(number: "12345").first_or_create!
+        }.to raise_error(Mongoid::Errors::Validations)
+      end
+    end
+
+    context "when passing a block" do
+
+      let(:account) do
+        Account.where(number: "12345").first_or_create! do |account|
+          account.name = "testing"
+        end
+      end
+
+      it "passes the block to the create call" do
+        expect(account.name).to eq("testing")
+      end
+
+      it "persists the new document" do
+        expect(account).to be_persisted
+      end
+    end
+
+    context "when the document is found" do
+
+      let!(:band) do
+        Band.with(safe: true).create!(name: "Depeche Mode")
+      end
+
+      let(:found) do
+        Band.where(name: "Depeche Mode").first_or_create!
+      end
+
+      it "returns the document" do
+        expect(found).to eq(band)
+      end
+    end
+
+    context "when the document is not found" do
+
+      let!(:band) do
+        Band.with(safe: true).create!(name: "Depeche Mode")
+      end
+
+      context "when attributes are provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create!(origin: "Essex")
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+
+        it "sets the additional attributes" do
+          expect(document.origin).to eq("Essex")
+        end
+      end
+
+      context "when attributes are not provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create!
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+      end
+
+      context "when a block is provided" do
+
+        let(:document) do
+          Band.where(name: "Tool").first_or_create! do |doc|
+            doc.active = false
+          end
+        end
+
+        it "returns a new document" do
+          expect(document.name).to eq("Tool")
+        end
+
+        it "returns a persisted document" do
+          expect(document).to be_persisted
+        end
+
+        it "yields to the block" do
+          expect(document.active).to be_false
+        end
+      end
+
+      context "when the criteria is complex" do
+
+        context "when the document is not found" do
+
+          let(:document) do
+            Band.in(name: [ "New Order" ]).first_or_create!(active: false)
+          end
+
+          it "returns a new document" do
+            expect(document.active).to be_false
+          end
+
+          it "returns a persisted document" do
+            expect(document).to be_persisted
+          end
+        end
+      end
+    end
+  end
 end